+++ js/data.js
@@ -1,0 +1,24 @@
+import {NAMES, COLLECTION_TEXT} from './util.js';
+
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+const createElement = () => {
+  const numberElement = getRandomInteger(0, 25);
+  return {
+    id: numberElement,
+    url: 'photos/' + numberElement + '.jpg',
+    description: 'Very long description of the photo',
+    likes: getRandomInteger(15, 200),
+    comments: new Array(6).fill().map(() => {
+      return {
+        id: String(numberElement) + getRandomInteger(0, 25),
+        avatar: 'img/avatar' + getRandomInteger(0, 6) + '.svg',
+        message: COLLECTION_TEXT[getRandomInteger(0, COLLECTION_TEXT.length)],
+        name: NAMES[getRandomInteger(0, NAMES.length)],
+      }
+    }),
+  };
+};
+
+new Array(25).fill(null).map(() => createElement());
+
